[Next Gen] refactor(codegen): split slot child-content emission into children
